Factor out get method test (#270)
* Factor out get method test

* Refactor to function
